fullstack/backend/requirements.txt
Flask
flask
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
scikit-learn
python-dotenv 
argon2-cffi
secure-smtplib
pandas